MSI-2557: Removed duplication of code. Only one service for SetShippingAddressesOnCartInterface
