--- app/code/Magento/InventoryInStorePickupQuoteGraphQl/etc/graphql/di.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <preference for="Magento\QuoteGraphQl\Model\Cart\SetShippingAddressesOnCartInterface" type="Magento\InventoryInStorePickupQuoteGraphQl\Model\Cart\SetShippingAddressesOnCart"/>
-</config>
